requirements.txt
# IDLock Backend - Requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Queue
redis==5.0.1
rq==1.16.0

# Settings
pydantic-settings==2.1.0

# HTTP Client
httpx>=0.28.1

# AI/ML - LLM Services
google-genai>=1.14.0
groq>=0.4.0

# AI/ML - Identity Extraction
numpy==1.26.3
insightface==0.7.3
onnxruntime>=1.17.0
opencv-python==4.9.0.80

# AI/ML - Vector Store (Local, FREE!)
faiss-cpu>=1.7.4

# AI/ML - Optional CLIP for style embeddings (comment out if not needed)
# torch>=2.0.0
# transformers>=4.35.0

# Image Processing
Pillow>=10.0.0

# Vector DB (Cloud option - needs API key)
pinecone-client==3.0.0

# Storage (Cloud options)
boto3==1.34.25  # AWS S3
google-cloud-storage==2.14.0  # Google Cloud Storage (for Cloud Run)
google-cloud-secret-manager==2.18.0  # Secret Manager (for Cloud Run)
google-cloud-aiplatform>=1.38.0  # Vertex AI (for LoRA training)

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Linting
ruff==0.1.14

